# packages/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: EquipChain Global Ltd API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /quotes:
    post:
      operationId: submitQuoteRequest
      tags: [quotes]
      summary: Submit quote request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteRequestInput"
      responses:
        "201":
          description: Submitted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"

    get:
      operationId: listQuoteRequests
      tags: [quotes]
      summary: List quote requests (admin)
      parameters:
        - name: page
          in: query
          schema:
            type: integer
            default: 1
        - name: limit
          in: query
          schema:
            type: integer
            default: 20
        - name: status
          in: query
          schema:
            type: string
            enum: [new, in_review, responded, closed]
        - name: search
          in: query
          schema:
            type: string
      responses:
        "200":
          description: List of quote requests
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteRequestList"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /quotes/{id}:
    get:
      operationId: getQuoteRequest
      tags: [quotes]
      summary: Get quote request by ID (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Quote request
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteRequest"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    delete:
      operationId: deleteQuoteRequest
      tags: [quotes]
      summary: Delete quote request (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/SuccessResponse"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /quotes/{id}/status:
    patch:
      operationId: updateQuoteStatus
      tags: [quotes]
      summary: Update quote request status (admin)
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/QuoteStatusInput"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/QuoteRequest"
        "401":
          description: Unauthorized
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  # Settings
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
    SuccessResponse:
      type: object
      required: [message]
      properties:
        message:
          type: string
...
    QuoteRequestInput:
      type: object
      required: [fullName, email, serviceType]
      properties:
        fullName:
          type: string
        companyName:
          type: string
        email:
          type: string
        phone:
          type: string
        serviceType:
          type: string
        otherService:
          type: string
        productService:
          type: string
        quantity:
          type: string
        specification:
          type: string
        deliveryTimeline:
          type: string
        sourcingType:
          type: string
        additionalMessage:
          type: string
        documentUrl:
          type: string
    QuoteRequest:
      type: object
      required: [id, fullName, email, serviceType, status, createdAt]
      properties:
        id:
          type: integer
        fullName:
          type: string
        companyName:
          type: ["string", "null"]
        email:
          type: string
        phone:
          type: ["string", "null"]
        serviceType:
          type: string
        otherService:
          type: ["string", "null"]
        productService:
          type: ["string", "null"]
        quantity:
          type: ["string", "null"]
        specification:
          type: ["string", "null"]
        deliveryTimeline:
          type: ["string", "null"]
        sourcingType:
          type: ["string", "null"]
        additionalMessage:
          type: ["string", "null"]
        documentUrl:
          type: ["string", "null"]
        status:
          type: string
          enum: [new, in_review, responded, closed]
        createdAt:
          type: string
    QuoteRequestList:
      type: object
      required: [requests, total, page, limit]
      properties:
        requests:
          type: array
          items:
            $ref: "#/components/schemas/QuoteRequest"
        total:
          type: integer
        page:
          type: integer
        limit:
          type: integer
    QuoteStatusInput:
      type: object
      required: [status]
      properties:
        status:
          type: string
          enum: [new, in_review, responded, closed]
...