# packages/api-zod/src/generated/types/quoteRequestInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * EquipChain Global Ltd API
 * OpenAPI spec version: 0.1.0
 */

export interface QuoteRequestInput {
  fullName: string;
  companyName?: string;
  email: string;
  phone?: string;
  serviceType: string;
  otherService?: string;
  productService?: string;
  quantity?: string;
  specification?: string;
  deliveryTimeline?: string;
  sourcingType?: string;
  additionalMessage?: string;
  documentUrl?: string;
}